Move NavigationExt to extensions package
