+++ app/src/main/java/com/infomaniak/mail/ui/main/settings/SettingsFragment.kt
@@ -37,8 +37,8 @@ import com.infomaniak.mail.data.LocalSettings
 import com.infomaniak.mail.data.models.FeatureFlag
 import com.infomaniak.mail.databinding.FragmentSettingsBinding
 import com.infomaniak.mail.ui.MainViewModel
-import com.infomaniak.mail.utils.animatedNavigation
-import com.infomaniak.mail.utils.launchSyncAutoConfigActivityForResult
+import com.infomaniak.mail.utils.extensions.animatedNavigation
+import com.infomaniak.mail.utils.extensions.launchSyncAutoConfigActivityForResult
 import com.infomaniak.mail.utils.extensions.observeNotNull
 import com.infomaniak.mail.utils.extensions.setSystemBarsColors
 import dagger.hilt.android.AndroidEntryPoint
