+++ app/src/main/java/com/infomaniak/mail/ui/main/settings/mailbox/MailboxSettingsFragment.kt
@@ -27,7 +27,7 @@ import com.infomaniak.lib.core.utils.safeBinding
 import com.infomaniak.mail.R
 import com.infomaniak.mail.databinding.FragmentMailboxSettingsBinding
 import com.infomaniak.mail.ui.main.settings.ItemSettingView
-import com.infomaniak.mail.utils.animatedNavigation
+import com.infomaniak.mail.utils.extensions.animatedNavigation
 import com.infomaniak.mail.utils.extensions.notYetImplemented
 import com.infomaniak.mail.utils.extensions.setSystemBarsColors
