+++ app/src/main/java/com/infomaniak/mail/ui/main/settings/send/SendSettingsFragment.kt
@@ -26,7 +26,7 @@ import com.infomaniak.lib.core.utils.safeBinding
 import com.infomaniak.mail.R
 import com.infomaniak.mail.data.LocalSettings
 import com.infomaniak.mail.databinding.FragmentSendSettingsBinding
-import com.infomaniak.mail.utils.animatedNavigation
+import com.infomaniak.mail.utils.extensions.animatedNavigation
 import com.infomaniak.mail.utils.extensions.notYetImplemented
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
