+++ app/src/main/java/com/infomaniak/mail/ui/main/user/SwitchUserFragment.kt
@@ -28,7 +28,7 @@ import com.infomaniak.lib.core.utils.safeBinding
 import com.infomaniak.mail.MatomoMail.trackAccountEvent
 import com.infomaniak.mail.databinding.FragmentSwitchUserBinding
 import com.infomaniak.mail.utils.AccountUtils
-import com.infomaniak.mail.utils.launchLoginActivity
+import com.infomaniak.mail.utils.extensions.launchLoginActivity
 import com.infomaniak.mail.utils.extensions.setSystemBarsColors
 import dagger.hilt.android.AndroidEntryPoint
