+++ app/src/main/java/com/infomaniak/mail/utils/extensions/NavigationExt.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.infomaniak.mail.utils
+package com.infomaniak.mail.utils.extensions
 
 import android.content.Context
 import android.content.Intent
@@ -38,7 +38,7 @@ import com.infomaniak.mail.ui.login.NoMailboxActivity
 import com.infomaniak.mail.ui.main.thread.actions.AttachmentActionsBottomSheetDialog
 import com.infomaniak.mail.ui.newMessage.NewMessageActivityArgs
 import com.infomaniak.mail.ui.noValidMailboxes.NoValidMailboxesActivity
-import com.infomaniak.mail.utils.extensions.AttachmentExt
+import com.infomaniak.mail.utils.AccountUtils
 import com.infomaniak.mail.utils.extensions.AttachmentExt.createDownloadDialogNavArgs
 
 fun getAnimatedNavOptions() = NavOptions
